requirements.txt
# NewERP - Manufacturing System
# Zależności Python
# Instalacja: pip install -r requirements.txt

# === WYMAGANE ===

# Supabase
supabase>=2.0.0

# Środowisko
python-dotenv>=1.0.0

# GUI
customtkinter>=5.2.0
Pillow>=10.0.0

# HTTP
requests>=2.31.0

# === DXF/CAD 2D - dla miniatur ===
ezdxf>=1.1.0
matplotlib>=3.7.0
numpy>=1.24.0

# === 3D CAD - dla miniatur STEP/IGES ===
# UWAGA: Instalacja VTK i CadQuery może zająć kilka minut
# Jeśli nie potrzebujesz miniatur 3D, możesz pominąć te pakiety

# VTK - renderowanie 3D
vtk>=9.2.0

# CadQuery - import STEP/IGES
cadquery>=2.3.0

# === OPCJONALNE ===

# Archiwa 7z
py7zr>=0.20.0

# Excel - do importu cenników
openpyxl>=3.1.0

# Raporty PDF
reportlab>=4.0.0

# Nesting - operacje na wielokątach (NFP)
pyclipper>=1.3.0

# Nesting - bin packing (FastNester)
rectpack>=0.2.2

# Nesting - geometria (buffer, intersects, contains)
shapely>=2.0.0

# PySide6 - GUI dla viewera 3D (osobna aplikacja)
# PySide6>=6.5.0

# Dokumenty Office
# python-docx>=0.8.11
# openpyxl>=3.1.0

# Development
# pytest>=7.0.0
# black>=23.0.0
